+++ CHANGELOG.rst
@@ -2,7 +2,7 @@
 Changelog for package rodney_sim_control
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-Unreleased
+0.7.0 (2019-05-02)
 ------------------
 * Changes so that simulation can run mission 1,2 and 4